Move NXdata views into DataViews, make them private
